+++ pet_product.rb
@@ -1,0 +1,42 @@
+# frozen_string_literal: true
+
+# Collecting info about pet product
+class PetProduct
+  attr_reader :product_link, :parser, :num
+
+  def initialize(product_link, num)
+    @parser = PetProductsParser.new(product_link)
+    @num = num
+  end
+
+  def index_num
+    puts num
+    num + 1
+  end
+
+  def name
+    parser.fetch_page.xpath("//h1[contains(@class, 'product_main_name')]").text
+  end
+
+  def weight_arr
+    parser.fetch_page.xpath("#{weight_and_price_fieldset}[1]").map(&:text)
+  end
+
+  def price_arr
+    parser.fetch_page.xpath("#{weight_and_price_fieldset}[last()]").map(&:text)
+  end
+
+  def image_link
+    parser.fetch_page.xpath("//img[contains(@id, 'bigpic')]/@src").text
+  end
+
+  def multiple?
+    weight_arr.length > 1 || price_arr.length > 1
+  end
+
+  private
+
+  def weight_and_price_fieldset
+    "//label[contains(@class, 'label_comb_price')]//span"
+  end
+end
